Script customization to allow frontend uploads

+++ assets/admin/js/admin-ajax-upload.js
@@ -1,15 +1,17 @@
 (function($){
 
-    var uploaded = false;
+    var uploaded  = false;
+    var ticket_id = false;
     Dropzone.autoDiscover = false;
 
     // Attach dropzone to every upload field
     $('.wpas-uploader-dropzone').each(function(i, e){
 
-        var id = $(e).attr('id');
+        var id    = $(e).attr('id');
+        ticket_id = (  $('#post_ID').length ) ? $('#post_ID').val() : $(this).data('ticket-id');
 
         $('#' + id).dropzone({ 
-            url: ajaxurl,
+            url: WPAS_AJAX.ajax_url,
             paramName: 'wpas_files',
             acceptedFiles : WPAS_AJAX.accept,
             maxFiles : WPAS_AJAX.max_files,
@@ -19,12 +21,12 @@
                 
                 this.on('sending', function(file, xhr, formData){
                     formData.append('action', 'wpas_upload_attachment');
-                    formData.append('ticket_id', $('#post_ID').val() );
+                    formData.append('ticket_id', ticket_id );
                     uploaded = true;
                 });
         
                 this.on('addedfile', function(e) {
-                    while (this.files.length > this.options.maxFiles) {
+                    if (this.files.length > this.options.maxFiles) {
                         this.removeFile(this.files[0]);
                         alert( WPAS_AJAX.exceeded );
                     }
@@ -38,14 +40,11 @@
 
                 $.ajax({
                     type: 'POST',
-                    url: ajaxurl,
+                    url: WPAS_AJAX.ajax_url,
                     data: {
                         action: 'wpas_delete_temp_attachment',
-                        parent_id: $('#post_ID').val(),
+                        ticket_id: ticket_id,
                         attachment: file.name
-                    },
-                    done: function(data){
-                        console.log('success: ' + data);
                     }
                 });
 
@@ -58,7 +57,7 @@
 
 
     // Reset uploaded var on reply submit
-    $('button[name="wpas_do"]').on('click', function(e){
+    $('button[name="wpas_do"], button[name="wpas-submit"]').on('click', function(e){
         uploaded = false;
     });
 
@@ -69,11 +68,11 @@
 
             $.ajax({
                 type: 'POST',
-                url: ajaxurl,
+                url: WPAS_AJAX.ajax_url,
                 async: false,
                 data: {
                     action: 'wpas_delete_temp_directory',
-                    parent_id: $('#post_ID').val(),
+                    ticket_id: ticket_id,
                 }
             });
         }
